requirements.txt
streamlit
python-dotenv
yt-dlp
opencv-python
numpy
google-generativeai
Pillow
fpdf2
markdown
markdownify